Fix type consistency: MasterHistory.id to Int, DAO return Int

+++ app/src/main/java/com/example/lotteryapp/data/MasterHistoryDao.kt
@@ -16,7 +16,7 @@ interface MasterHistoryDao {
     suspend fun getById(id: Int): MasterHistory?
     
     @Insert
-    suspend fun insert(master: MasterHistory): Long
+    suspend fun insert(master: MasterHistory): Int
     
     @Update
     suspend fun update(master: MasterHistory)
